Convert Point to use double instead of float
Because Justin said so.

- Emmie Jones

+++ src/main/include/util/point.h
@@ -8,16 +8,16 @@ static const double NEAR_ZERO_METERS = 0.005;
 
 class Point {
     public:
-        Point(float x, float y, float z) : x(x), y(y), z(z) {}
+        Point(double x, double y, double z) : x(x), y(y), z(z) {}
         Point() { // default constructor
             x = 0;
             y = 0;
             z = 0;
         } // default constructor
 
-        float x;
-        float y;
-        float z;
+        double x;
+        double y;
+        double z;
 
 
     Vector operator- (Point const & rhs) const {
